#7 - DuyHD7 - Implement Order Service
Update create client order and integrate paypal

+++ order-service/src/main/java/com/dcode/order_service/dto/order/request/OrderLineRequest.java
@@ -1,10 +1,12 @@
 package com.dcode.order_service.dto.order.request;
 
+import java.math.BigDecimal;
+
 public record OrderLineRequest(
         String orderId,
         String productId,
-        Double quantity,
-        Double trackingPrice,
+        Integer quantity,
+        BigDecimal trackingPrice,
         String variantId,
         String paintId,
         String wallpaperId,
